requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.1
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.3
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
decouple==0.0.7
defusedxml==0.6.0
dj-database-url==0.5.0
dj-stripe==2.3.0
Django==3.0.8
django-allauth==0.43.0
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-money==1.2.1
django-paypal==1.0.0
django-rest-framework==0.1.0
django-richtextfield==1.6
djangorestframework==3.12.2
docutils==0.16
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.9
incremental==17.5.0
jsonfield==3.1.0
Kivy-Garden==0.1.4
msgpack==1.0.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.55.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
whitenoise==5.2.0
zope.interface==5.2.0
